pyproject.toml
[tool.poetry]
name = "ytmarkdl"
version = "2.0"
description = "Download Youtube music from Chrome bookmarks"
authors = ["Crossoufire"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
eyed3 = "^0.9.7"
rich = "^13.9.4"
spotipy = "^2.24.0"
mutagen = "^1.47.0"
yt-dlp = "^2024.11.4"
python-dotenv = "^1.0.1"
python = "^3.10 || ^3.11 || ^3.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
